refactor: bound Turso spec publication module
